+++ scrapper/src/main/java/edu/java/configuration/ApplicationConfig.java
@@ -1,6 +1,5 @@
 package edu.java.configuration;
 
-import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import java.time.Duration;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -13,9 +12,7 @@ public record ApplicationConfig(
     @Bean
     @NotNull
     Scheduler scheduler,
-    @NotEmpty
     String githubBaseUrl,
-    @NotEmpty
     String stackoverflowBaseUrl
 ) {
     public record Scheduler(boolean enable, @NotNull Duration interval, @NotNull Duration forceCheckDelay) {
