+++ scrapper/src/main/java/edu/java/configuration/GitHubClientConfig.java
@@ -1,7 +1,6 @@
-package edu.java.github;
+package edu.java.configuration;
 
 import edu.java.WebClientErrorHandler;
-import edu.java.configuration.ApplicationConfig;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.reactive.function.client.WebClient;
