+++ scrapper/src/main/resources/application.yml
@@ -4,8 +4,8 @@ app:
     interval: 10s
     force-check-delay: 10s
 
-  github_base_url: ${GITHUB_BASE_URL}
-  stackoverflow_base_url: ${STACK_OVERFLOW_BASE_URL}
+  github_base_url: ${GITHUB_BASE_URL:https://api.github.com/}
+  stackoverflow_base_url: ${STACK_OVERFLOW_BASE_URL:https://api.stackexchange.com/2.3/}
 
 spring:
   application:
