+++ scrapper/src/test/java/edu/scrapper/github/GitHubClientTest.java
@@ -2,43 +2,23 @@ package edu.scrapper.github;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.client.WireMock;
-import edu.java.github.GitHubClient;
 import edu.java.github.GitHubClientImpl;
 import edu.java.github.RepositoryResponse;
+import java.io.File;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.time.OffsetDateTime;
+import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.web.reactive.function.client.WebClient;
-import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.options;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-@Import({GitHubClientTest.TestConfig.class})
+@Import({GitHubTestConfiguration.class})
 @ExtendWith(SpringExtension.class)
 public class GitHubClientTest {
-    @TestConfiguration
-    static class TestConfig {
-        @Bean
-        public WireMockServer wireMockServer() {
-            WireMockServer wireMockServer = new WireMockServer(options().dynamicPort());
-            wireMockServer.start();
-            return wireMockServer;
-        }
-
-        @Bean
-        public WebClient webClient(WireMockServer server) {
-            return WebClient.builder().baseUrl(server.baseUrl()).build();
-        }
-
-        @Bean
-        public GitHubClient gitHubClient(WebClient webClient) {
-            return new GitHubClientImpl(webClient);
-        }
-    }
 
     @Autowired
     private WireMockServer wireMockServer;
@@ -56,17 +36,17 @@ public class GitHubClientTest {
     }
 
     @Test
-    public void testGetRepository() {
+    public void testGetRepository() throws IOException {
+        String okResponse = FileUtils.readFileToString(
+            new File("src/test/resources/github/github_ok_response.json"),
+            StandardCharsets.UTF_8
+        );
         wireMockServer.stubFor(WireMock.get(WireMock.urlPathEqualTo("/repos/testOwner/testRepo"))
             .willReturn(WireMock.aResponse()
                 .withHeader("Content-Type", "application/json")
-                .withBody(
-                    "{\"id\": \"1\", " +
-                        "\"name\": \"testRepo\", " +
-                        "\"description\": \"Test repository\", " +
-                        "\"html_url\": \"https://github.com/testOwner/testRepo\", " +
-                        "\"updated_at\": \"2024-02-27T12:00:00Z\"}"
-                )));
+                .withBody(okResponse)
+            )
+        );
 
         RepositoryResponse repositoryResponse = gitHubClient.getRepository("https://github.com/testOwner/testRepo");
         RepositoryResponse expectedResponse = new RepositoryResponse(
