+++ scrapper/src/test/resources/github/github_ok_response.json
@@ -1,0 +1,7 @@
+{
+    "id": 1,
+    "name": "testRepo",
+    "description": "Test repository",
+    "html_url": "https://github.com/testOwner/testRepo",
+    "updated_at": "2024-02-27T12:00:00Z"
+}
