+++ scrapper/src/test/resources/stackoverflow/stackoverflow_ok_response.json
@@ -1,0 +1,13 @@
+{
+    "items": [
+        {
+            "is_answered": true,
+            "answer_count": 2,
+            "last_activity_date": 1709035200,
+            "creation_date": 1256799465,
+            "question_id": 1,
+            "link": "https://stackoverflow.com/questions/1/test-question",
+            "title": "Test question"
+        }
+    ]
+}
